moved get_noise() method into img2img class
